Inserindo novo controller e documentacao

+++ src/main/java/br/com/questions/repository/VoteAnswerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VoteAnswerRepository extends JpaRepository<VoteAnswer, Integer>{
+public interface VoteAnswerRepository extends JpaRepository<VoteAnswer, Long>{
 }
